Fix: Implemented reviewer feedback

+++ bundles/CoreBundle/Migrations/Version20221222181745.php
@@ -34,9 +34,25 @@ final class Version20221222181745 extends AbstractMigration
             $theTableGroups = current($tableGroups);
             $tableArray = explode('_', $theTableGroups);
             $tableNumber = end($tableArray);
+            if(!is_numeric($tableNumber)){
+                continue;
+            }
+
+            $this->addSql("DELETE `object_classificationstore_groups_{$tableNumber}`
+            FROM `object_classificationstore_groups_{$tableNumber}`
+            LEFT JOIN `classificationstore_groups` ON object_classificationstore_groups_{$tableNumber}.groupId = classificationstore_groups.id
+            WHERE classificationstore_groups.id = NULL;");
+
+            $this->addSql("DELETE `object_classificationstore_data_{$tableNumber}`
+            FROM `object_classificationstore_data_{$tableNumber}`
+            LEFT JOIN `$theTableGroups` ON object_classificationstore_data_{$tableNumber}.o_id = $theTableGroups.o_id AND
+            object_classificationstore_data_{$tableNumber}.fieldname = $theTableGroups.fieldname AND
+            object_classificationstore_data_{$tableNumber}.groupId = $theTableGroups.groupId
+            WHERE $theTableGroups.o_id = NULL AND  $theTableGroups.fieldname = NULL AND $theTableGroups.groupId = NULL;");
+
             $this->addSql("ALTER TABLE `$theTableGroups` MODIFY COLUMN groupId INT(11) UNSIGNED NOT NULL;");
             $this->addSql("ALTER TABLE `$theTableGroups`
-            ADD CONSTRAINT `fk_object_classificationstore_groups_{$tableNumber}__groupId`FOREIGN KEY (`groupId`)
+            ADD CONSTRAINT `fk_object_classificationstore_groups_{$tableNumber}__groupId` FOREIGN KEY (`groupId`)
             REFERENCES `classificationstore_groups` (`id`)
             ON DELETE CASCADE;");
 
@@ -44,7 +60,7 @@ final class Version20221222181745 extends AbstractMigration
             $this->addSql("ALTER TABLE `$theTableData` MODIFY COLUMN groupId INT(11) UNSIGNED NOT NULL;");
             $this->addSql("CREATE INDEX `groupKeys` ON `$theTableData` (`o_id`, `fieldname`, `groupId`);");
             $this->addSql("ALTER TABLE `$theTableData`
-            ADD CONSTRAINT `fk_object_classificationstore_data_{$tableNumber}__o_id__fieldname__groupId`FOREIGN KEY (`o_id`, `fieldname`, `groupId`)
+            ADD CONSTRAINT `fk_object_classificationstore_data_{$tableNumber}__o_id__fieldname__groupId` FOREIGN KEY (`o_id`, `fieldname`, `groupId`)
             REFERENCES `$theTableGroups` (`o_id`, `fieldname`, `groupId`)
             ON DELETE CASCADE;");
         }
@@ -57,13 +73,19 @@ final class Version20221222181745 extends AbstractMigration
             $theTableGroups = current($theTableGroups);
             $tableArray = explode('_', $theTableGroups);
             $tableNumber = end($tableArray);
+            if(!is_numeric($tableNumber)){
+                continue;
+            }
+
             $this->addSql("ALTER TABLE `$theTableGroups` DROP FOREIGN KEY `fk_object_classificationstore_groups_{$tableNumber}__groupId`;");
-            $this->addSql("ALTER TABLE `$theTableGroups` MODIFY COLUMN groupId BIGINT(20) NOT NULL;");
 
             $theTableData = "object_classificationstore_data_$tableNumber";
             $this->addSql("ALTER TABLE `$theTableData` DROP FOREIGN KEY `fk_object_classificationstore_data_{$tableNumber}__o_id__fieldname__groupId`;");
             $this->addSql("ALTER TABLE `$theTableData` DROP INDEX `groupKeys`;");
+
             $this->addSql("ALTER TABLE `$theTableData` MODIFY COLUMN groupId BIGINT(20) NOT NULL;");
+            $this->addSql("ALTER TABLE `$theTableGroups` MODIFY COLUMN groupId BIGINT(20) NOT NULL;");
+
         }
     }
